[kostalinverter] Fixed package structure
The internal package was placed before the binding specific package, and it should be the other way around.
