+++ bundles/org.openhab.binding.kostalinverter/src/main/java/org/openhab/binding/kostalinverter/internal/firstgeneration/WebscrapeHandler.java
@@ -10,7 +10,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.openhab.binding.internal.kostal.inverter.firstgeneration;
+package org.openhab.binding.kostalinverter.internal.firstgeneration;
 
 import java.io.IOException;
 import java.math.BigDecimal;
